It works! Well, sort of.
